requirements.txt
boto3==1.28.40
botocore==1.31.40
jmespath==1.0.1
names==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.16
